+++ tests/Australia/NorthernTerritory/PicnicDayTest.php
@@ -10,7 +10,7 @@
  * @author Sacha Telgenhof <me@sachatelgenhof.com>
  */
 
-namespace Yasumi\tests\Australia\NT;
+namespace Yasumi\tests\Australia\NorthernTerritory;
 
 use DateTime;
 use DateTimeZone;
@@ -20,22 +20,17 @@ use Yasumi\Holiday;
 use Yasumi\tests\YasumiTestCaseInterface;
 
 /**
- * Class for testing Queen's Birthday in NT (Australia)..
+ * Class for testing Picnic Day in Northern Territory (Australia).
  */
-class QueensBirthdayTest extends NTBaseTestCase implements YasumiTestCaseInterface
+class PicnicDayTest extends NorthernTerritoryBaseTestCase implements YasumiTestCaseInterface
 {
     /**
      * The name of the holiday
      */
-    public const HOLIDAY = 'queensBirthday';
+    public const HOLIDAY = 'picnicDay';
 
     /**
-     * The year in which the holiday was first established
-     */
-    public const ESTABLISHMENT_YEAR = 1950;
-
-    /**
-     * Tests Queen's Birthday
+     * Tests Picnic Day
      *
      * @dataProvider HolidayDataProvider
      *
@@ -63,17 +58,17 @@ class QueensBirthdayTest extends NTBaseTestCase implements YasumiTestCaseInterfa
     public function HolidayDataProvider(): array
     {
         return [
-            [2010, '2010-06-14'],
-            [2011, '2011-06-13'],
-            [2012, '2012-06-11'],
-            [2013, '2013-06-10'],
-            [2014, '2014-06-09'],
-            [2015, '2015-06-08'],
-            [2016, '2016-06-13'],
-            [2017, '2017-06-12'],
-            [2018, '2018-06-11'],
-            [2019, '2019-06-10'],
-            [2020, '2020-06-08'],
+            [2010, '2010-08-02'],
+            [2011, '2011-08-01'],
+            [2012, '2012-08-06'],
+            [2013, '2013-08-05'],
+            [2014, '2014-08-04'],
+            [2015, '2015-08-03'],
+            [2016, '2016-08-01'],
+            [2017, '2017-08-07'],
+            [2018, '2018-08-06'],
+            [2019, '2019-08-05'],
+            [2020, '2020-08-03'],
         ];
     }
 
@@ -86,8 +81,8 @@ class QueensBirthdayTest extends NTBaseTestCase implements YasumiTestCaseInterfa
         $this->assertTranslatedHolidayName(
             $this->region,
             self::HOLIDAY,
-            $this->generateRandomYear(self::ESTABLISHMENT_YEAR),
-            [self::LOCALE => 'Queen’s Birthday']
+            $this->generateRandomYear(1990),
+            [self::LOCALE => 'Picnic Day']
         );
     }
 
@@ -97,11 +92,6 @@ class QueensBirthdayTest extends NTBaseTestCase implements YasumiTestCaseInterfa
      */
     public function testHolidayType(): void
     {
-        $this->assertHolidayType(
-            $this->region,
-            self::HOLIDAY,
-            $this->generateRandomYear(self::ESTABLISHMENT_YEAR, 2100),
-            Holiday::TYPE_OFFICIAL
-        );
+        $this->assertHolidayType($this->region, self::HOLIDAY, $this->generateRandomYear(1990), Holiday::TYPE_OFFICIAL);
     }
 }
